Added Merge Projects to File Menu
Merging projects is now supported.  If two or more users are processing
data from the same test, the projects can be combined into one using the
'Merge Projects' utility in the 'File' menu. Preference is given to the
project which was opened when the File Path was specified.  In other
words, the utility only imports data from the file selected when Merge
Projects runs which was NOT contained in the original file.  Therefore,
no data from the original project will be over written. Only missing
data will be imported.